microservice_bundle/requirements.txt

fastapi==0.109.2
uvicorn[standard]==0.27.0
google-generativeai>=0.3.2
python-multipart==0.0.7
psycopg2-binary==2.9.9
supabase==2.3.4
pydantic==2.6.0
pydantic-settings==2.1.0
requests==2.31.0
python-dotenv==1.0.1
